+++ fabric/src/main/java/dev/aika/taczjs/fabric/events/crafting/legacy/RecipeLoadEndEvent.java
@@ -1,4 +1,0 @@
-package dev.aika.taczjs.fabric.events.crafting.legacy;
-
-public class RecipeLoadEndEvent extends RecipeLoadBeginEvent {
-}