+++ neoforge/src/main/java/dev/aika/taczjs/neoforge/events/crafting/legacy/RecipeLoadEndEvent.java
@@ -1,0 +1,4 @@
+package dev.aika.taczjs.neoforge.events.crafting.legacy;
+
+public class RecipeLoadEndEvent extends RecipeLoadBeginEvent {
+}